+++ src/Core/Dfe.Complete.Application/Projects/Interfaces/IProjectWriteRepository.cs
@@ -1,0 +1,9 @@
+﻿using Dfe.Complete.Domain.Entities;
+
+namespace Dfe.Complete.Application.Projects.Interfaces
+{
+    public interface IProjectWriteRepository
+    {
+        Task UpdateProjectAsync(Project project, CancellationToken cancellationToken);
+    }
+}